+++ src/entities/particle_system.h
@@ -20,10 +20,10 @@ class ParticleEngine : public Entity {
 
   /*! \brief Determine how many particles should be emitted based on the
    * firerate, time since last update, and number of alive particles. */
-  int queued_shots(float time_since);
+  int queued_shots(double time_since);
 
   /*! \brief Create and emit particles based on how much time passed.. */
-  void create_new_particles(float time);
+  void create_new_particles(double time);
 
   /*! \brief Emit as many particles as possible */
   void emit_particle(int num_particles);
@@ -33,13 +33,13 @@ class ParticleEngine : public Entity {
 
   /*! \brief Apply simulation to every live particle, detect and remove dead
    * particles, and create new particles */
-  void simulate_particles(float time);
+  void simulate_particles(double time);
 
  public:
   // EMITTER
   int max_particles =
       15000;  //< Maximum number of particles alive at any given time
-  float fire_rate =
+  double fire_rate =
       0.001;  //< Minimum delay between the creation of each particle
 
   int particles_per_second = 1000;
